+++ src/PromptingTools.jl
@@ -117,6 +117,7 @@ include("llm_ollama_managed.jl")
 include("llm_ollama.jl")
 include("llm_google.jl")
 include("llm_anthropic.jl")
+include("llm_langdock.jl")
 include("llm_sharegpt.jl")
 include("llm_tracer.jl")